refactored code (#1)
